#4756: Rework dispatch-agnostic gtest fixtures to allow for more types
